general endpoint for login
